# optimized EnumSet::count()