## Bench

### Before
```
$ php bench.php 
empty set count(): 0.14792680740356ms
full set count() : 0.13878393173218ms
```

### After
```
$ php bench.php 
empty set count(): 0.0060579776763916ms
full set count() : 0.10427093505859ms
```